refactor code to make adding more games easier
