+++ src/global.d.ts
@@ -18,3 +18,8 @@ declare module '*.gif' {
     const value: string;
     export default value;
 }
+
+declare module '*.webp' {
+    const value: string;
+    export default value;
+}
